added new class WithinExamples.cs (extracted tests from ApiExamples.cs)
